Integrate with Deeplift
I now get a scores vector with an entry for each of the columns. Need to
look at which columns get which scores.

Then dick with scaling, especially since remove BatchNorm makes this
perform terribly. Also, use_bias is necessary even with BatchNorm.

Deeplift only works with tfv1, so had to change a bunch of little stupid
things.